imx6-mmdc: Work around ERR050070
The MPWLGCR registers clear the error bits before software can read them,
so rely on the MPWLHWERR registers for error reporting instead.

This errata was announced in 2019 but it seems to apply to all chip revisions.

U-Boot contains a different workaround where it instead checks the calibration
data for the result 0x001F001F (maximum delay) and flag that as a failure.
I can't find the origin of this workaround but I first saw it in the Novena
source code, though I asked Sean Cross and he suggested it was from Freescale.

While we're at it, fix the comment implying this code only checks PHY0 in x32
configuration. This is wrong and misleading.

Signed-off-by: John Watts <contact@jookia.org>
Link: https://lore.barebox.org/20230124174115.1509216-1-contact@jookia.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>