spi: Fix probing SPI drivers with no cs-gpios
of_gpio_named_count returns a negative value on error but this
is discarded and cast to a u16, making error handling impossible.

With debug logging enabled this effectively halts booting so the board can
print an error over serial 65534 times.

Introduce of_gpio_count_csgpios which returns 0 in the case of an error
rather than a negative value.

Signed-off-by: John Watts <contact@jookia.org>
Link: https://lore.barebox.org/20230124200554.1701687-1-contact@jookia.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>