i2c: implement detect callback for virtual adapter device as well
For MMCs, we already support detect on both the hardware device and the
mmcX virtual device. Let's do the same for i2c, so users have the option
to do `detect i2c0` instead of `detect 30a30000.i2c@30a30000.of`.

`detect -a` still works as expected, as the detect callback is a no-op
if everything on the bus is already registered.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20230125075323.1759934-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>