ARM: i.MX6: Specify OCRAM base address and size
The i.MX6 includes some on-chip RAM: 128KiB on most variants, with 256KiB on
the Dual and Quad variants.

This region is where the first stage of Barebox gets loaded if RAM
initialization isn't hard coded using DCD information.

Add the base address and size so it can be used to calculate where to put the
stack at boot.

Signed-off-by: John Watts <contact@jookia.org>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20230126190258.110423-1-contact@jookia.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>