ddr_dimms: Check spd->mem_type before computing parameters
Checking the mem_type here as well as the CRC means developers can
just read the SPD and pass it directly to parameter computation.

This makes it so developers can rely fully on parameters for checking
if RAM is compatible with their board and not need to worry about reading
SPD data themselves.

Signed-off-by: John Watts <contact@jookia.org>
Link: https://lore.barebox.org/20230201053334.971854-1-contact@jookia.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>