I2C: i.MX: early: Use a custom delay on i.MX6
The i.MX early I2C code requires waiting for the controller to settle
after configuration. This is currently done using udelay which is
supported on ARMv8 but not on ARMv7.

For the i.MX6 we will have to use a custom delay. This was only
tested on the i.MX6Q but should work on all other i.MX6 chips.

Signed-off-by: John Watts <contact@jookia.org>
Link: https://lore.barebox.org/20230201094435.1228362-1-contact@jookia.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>