ARM: novena: Use DDR3 information from SPD EEPROM
The Novena supports swappable memory which means we must query the memory
for configuration details. This patch uses values stored in the SPD EEPROM
to configure the i.MX6 memory controller.

If the SPD can't be read, the default configuration for a 4GB stick is
used. This is the same type of stick that comes with the board.

Calibration for the installed stick is run unconditionally.

Verification that this works was done by comparing MMDC registers before and
after and carefully reading other people's source code.

Signed-off-by: John Watts <contact@jookia.org>
Reviewed-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20230202075741.2350028-5-contact@jookia.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>