fs: nfs: Error if a port is unregistered
If a server doesn't provide a service the port lookup will return 0.

Check for this, return an appropriate error code and give a hint as to
why this might be happening.

This was tested by running NFS with UDP disabled, which seems to be
the default in Arch at least.

Signed-off-by: John Watts <contact@jookia.org>
Link: https://lore.barebox.org/20230202112024.2458085-1-contact@jookia.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>