RISC-V: Makefile: cleanup layout
The cflag and cppflag handling is rather complicated for this small
Makefile. The code uses two helper variables riscv-cflags-y cflags-y
which are assigned later to PBL_CPPFLAGS, KBUILD_CPPFLAGS and
KBUILD_CFLAGS.

Remove this complexity by assigning the flags directly to
KBUILD_CPPFLAGS. The variable is used during pbl/barebox compile and
during cpp process.

While on it move the KBUILD_IMAGE to the KBUILD_BINARY to bundle the
usage.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20230202132612.691206-1-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>