ARM: i.MX8MP-EVK: increase VDD_ARM to OD voltage
The Linux kernel or whatever is started from Barebox might switch the
ARM frequency to OD level, without first reprogramming the PMIC as
required. This might lead to system instability. To avoid this,
increase VDD_ARM to OD level. When the kernel handles the PMIC properly
it will drop the voltage back to ND level when appropriate.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Link: https://lore.barebox.org/20230206205928.627270-1-l.stach@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>