ratp: Increase the initial RTO to 200ms
The initial value for RTO is 100ms which might be a bit low. From the
RFC916 the RTO is expected to have a lower and upper bound but values
are not specified.  The RFC916 also define the calculation of the RTO
to be somewhere between 1.3 to 2.0 times the SRTT (which is currently
defined to 100ms).  Thus I propose to set the initial value of RTO to
200ms, to be 2.0 times greater than the initial SRTT.

Moreover, the current runtime calculation for RTO is done in the
function ratp_msg_done and has lower bound of 200ms:
    ri->srtt = (alpha * ri->srtt + (10 - alpha) * rtt) / 10;
    ri->rto = max(200, beta * ri->srtt / 10);

Signed-off-by: Jules Maselbas <jmaselbas@kalray.eu>
Link: https://lore.barebox.org/20230207162055.10050-5-jmaselbas@kalray.eu
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>