Revert "sandbox: fix linker warnings with newer ld"
cc-option calls gcc with -c, so doesn't link internally. It is not
suitable for testing linker flags.
We use gcc for linking, so we need -Wl, for passing linker flags.
ld-option doesn't work either, because it expects and outputs
linker options directly without -Wl.
There is no easy solution to this, so revert the patch for now.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>