ARM: IMX8MP: add initial support for Variscite DT8MCustomBoard with iMX8MP
The Variscite DT8MCustomBoard is an eval board for the several Variscite
SOMs of their "DART" product line. This commit adds support for that
baseboard in combination with a DART-MX8M-PLUS SOM. The commit contains
an adapted version of the vendors device tree [1] and the vendors DDR
timings, taken from Variscite's public U-Boot sources [2].

Both files have been slightly changed to integrate well with barebox.

The boardsupport added with this commit does not yet support every
feature of the DT8MCustomBoard. Yet it already supports all basic
necessities to make use of barebox.

[1] https://github.com/varigit/linux-imx/tree/3f94f35bda827e8aa06beadb10c77358cfb6dad9
[2] https://github.com/varigit/uboot-imx/tree/7cad2ff68a508c71c572151a85bc786711bab969

Signed-off-by: Michael Kopfensteiner <michael.kopfensteiner@vahle.com>
Link: https://lore.barebox.org/20230223153601.59357-2-michael.kopfensteiner@vahle.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>