pm_domains: Enable explicitly when we have power-domain providers
Many SoCs provide power-domains in their device trees. In barebox
we most of the time do not have drivers for the power-domains. This
is no problem for most of the time as the necessary power domains
are either default on or enabled during early SoC init.
With multi-arch support it can happen that we compile in SoCs for
which we do not have power-domain drivers and other SoCs where we
need a power-domain driver. The example I stumbled upon is the
i.MX7 on the one hand which has a driver and OMAP on the other hand
which has power-domains described in the device tree for which
we do not have (and do not currently need) a driver.

To be able to handle both situations let's enable pm_domain support
explicitly on SoCs we know we have a driver for.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>