serial: ns16550: Fix device-tree probe and add proper teardown
This fix a new issue introduced by a previous patch that fixes a
previous issue.

The previous issue is related to the memory resource not being released
in case the probe fails or is deferred, hence if deferred the probe will
always fail to request the same memory resource. The proposed solution
was to move the resource request after the clock initialization which
was causing the driver to be deferred.

This introduced a new issue when using information from the device-tree,
the function ns16550_probe_dt() cannot easily be moved around as it:
 - adds offset to mmiobase, thus need to be called after ns16550_init_iomem()
 - sets the clock frequency, thus need to be called before clock init
 - sets {read,write}_reg callbacks using the width read from the device-tree,
   and is expected to override previous callbacks sets when calling
   ns16550_init_{iomem,ioport}

To fix this, both ns16550_init_iomem() and ns16550_init_ioport() functions
are fused and return the requested region. In case of an error during the
probe the resource is be properly released and can requested again if needed.

Fixes b32172c2f9 ("serial: ns16550: move iomem/ioport init after clock init")

Signed-off-by: Jules Maselbas <jmaselbas@kalray.eu>
Link: https://lore.barebox.org/20230303135253.15500-1-jmaselbas@kalray.eu
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>