common: elf: use proper field to get segment memory size
For a program header, the value of 'p_filesz' may be less than the one
of 'p_memsz', in which case the leftover amount of space should be
filled with zeroes. This is done correctly when loading elf to memory.
However, when we are requesting memory beforehand we do pass the value
of 'p_filesz', instead of 'p_memsz', to 'request_sdram_region',
resulting in potentially smaller amount of memory to be reserved.

As 'p_memsz' is always greater or equal to 'p_filesz', use only the
former for checking memory requirements in 'request_elf_segment'.

Signed-off-by: Denis Orlov <denorl2009@gmail.com>
Link: https://lore.barebox.org/20230315101126.4101087-1-denorl2009@gmail.com
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>