ARM: cleanup 32bit/64bit support in Kconfig
The decision whether to build a 32bit or 64bit barebox has to be made
first before anything else, so this makes CONFIG_64BIT a toplevel option
without any further dependencies.

With this patch we will only present the SoCs/boards which are actually
supported by the selected code model in Kconfig.

Without this patch it was often possible to select 32bit boards on a
64bit build or vice versa, which resulted in a broken build.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>