ARM: Rockchip: Do not pass device tree to TF-A
The downstream TF-A doesn't cope with our device tree when
CONFIG_OF_OVERLAY_LIVE is enabled, supposedly because it is
too big for some reason. Otherwise it doesn't have any visible
effect if we pass a device tree or not, except that the TF-A
fills in the ethernet MAC address into the device tree.
The upstream TF-A doesn't use the device tree at all.

Pass NULL for now until we have a good reason to pass a real
device tree.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>