ARM: dts: i.MX8MP: snps,dis-u2-freeclk-exists-quirk
Linux v6.1 adds a snps,gfladj-refclk-lpm-sel-quirk property to the
i.MX8MP DWC3 device tree nodes, which is a superset of the existing
snps,dis-u2-freeclk-exists-quirk property. For this reason, the old
property was dropped breaking some usages of barebox i.MX8MP DWC3.

The proper fix would be supporting this new property, which is already
in next as part of the DWC3 driver sync with the kernel.

Until that lands in master though, let's add back the removed property
to fix the regression. Expectation is that we can revert this commit
as part of v2023.05.0.

Link: https://github.com/saschahauer/barebox/issues/19
Reported-by: Hans Christian Lonstad <hcl@datarespons.com>
Fixes: 32e2176ba050 ("dts: update to v6.1-rc1")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230330113844.3343865-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>