RISC-V: configs: rv64i_defconfig: disable SBI serial
On Virt64, enabling SBI serial leads to doubled output, because both
OpenSBI and barebox use the same serial for output and all consoles
are enabled by default.

Given that SBI PUTCHAR is no longer part of newer SBI specs, we can
just disable the sbi console driver in the defconfig to work around
this.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230411071436.1630752-9-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>