Kbuild: export CROSS_PKG_CONFIG
CROSS_PKG_CONFIG is referenced in both Kconfig and Makefiles. To be
visibile everywhere, its default needs to be exported, but it wasn't.

This wasn't noticed, because usual usage was:

  CROSS_PKG_CONFIG=some-pkg-config make scripts

where the shell took care to export the variable. On NixOS, cross
pkg-config is actually called ${CROSS_COMPILE}pkg-config and it didn't
work out of the box. Fix this.

Fixes: 1c1198a3f252 ("scripts: allow building USB loader tools for target as well")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20230411071436.1630752-12-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>